Paranoia.

+++ install.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 INSTALL_DIR=$HOME/bin
-BASH_ALIAS_FILE=$HOME/.bashrc
+ALIAS_FILE=$HOME/.bashrc
 
 Usage(){
 	echo "Usage: $(basename $0) [options]"
@@ -9,22 +9,24 @@ Usage(){
 	echo -e " -h, --help\tdisplays this message and exits"
 	echo -e " --install-dir=/path/to/dir [Default: $INSTALL_DIR]
 		Uses path provided as home directory for todo.py"
-	echo -e " --alias-file=/path/to/file [Default: $BASH_ALIAS_FILE]
+	echo -e " --alias-file=/path/to/file [Default: $ALIAS_FILE]
 		Uses file to store alias for \$INSTALL_DIR/todo.py"
 	exit
 }
 
-set -- $(getopt -l "help,install-dir::,alias-file::" "h" "$@")
+set $(getopt -l "help,install-dir:,alias-file:" "h" "$@")
+
+[[ $? -ne 0 ]] || echo ; Usage ; exit  # If you fail to give me parameters
+                        # I'll fail to install it for you. Simple as that.
 
 while [[ $# -gt 0 ]] ; do 
 	case "$1" in
 		"-h" | "--help" ) Usage
 			;;
-		"--install-dir" ) shift; INSTALL_DIR=$1
+		"--install-dir" ) shift; [[ ! -z "$1" ]] && INSTALL_DIR=$1 || exit
 			;;
-		"--alias-file" ) shift; BASH_ALIAS_FILE=${1:1:${#1}-2} #[1]
-			;;
-		"--" ) break
+		"--alias-file" ) shift; [[ ! -z "$1" ]] && ALIAS_FILE=${1:1:${#1}-2} || exit
+            #[1]
 			;;
 		* ) shift
 	esac
@@ -35,16 +37,20 @@ prog="[""$(basename $0)""] "
 [[ -d $INSTALL_DIR ]] || mkdir -p $INSTALL_DIR
 echo $prog"$INSTALL_DIR exists."
 
-[[ -s $BASH_ALIAS_FILE ]] || echo "# Bash RC File" >> $BASH_ALIAS_FILE
-echo $prog"$BASH_ALIAS_FILE exists."
+[[ -s $ALIAS_FILE ]] || echo "# Bash RC File" >> $ALIAS_FILE
+echo $prog"$ALIAS_FILE exists."
 ## Believe it or not, >> is faster than >.
 
 echo $prog"Copying todo.py to $INSTALL_DIR/todo.py"
-cp ./todo.py $INSTALL_DIR
+if [[ ! -f $INSTALL_DIR/todo.py ]] ; then
+    cp ./todo.py $INSTALL_DIR
+else
+    cp -u ./todo.py $INSTALL_DIR
+fi
 
 ## Establish alias
 pre="\n\n#Alias for todo.py\n"
-if grep -q "todo.sh" "$BASH_ALIAS_FILE" ; then
+if grep -q -e "todo.sh" -e "todo.py" "$ALIAS_FILE" ; then
 	alias="tpy"
 else
 	alias="t"
@@ -52,9 +58,9 @@ fi
 
 alias_rc=$pre"alias "$alias"='$INSTALL_DIR/todo.py'\n"
 
-echo -e $alias_rc >> $BASH_ALIAS_FILE
-echo $prog"Alias '$alias' added to $BASH_ALIAS_FILE."
-echo $prog"To use alias, please run \`source $BASH_ALIAS_FILE\`."
+echo -e $alias_rc >> $ALIAS_FILE
+echo $prog"Alias '$alias' added to $ALIAS_FILE."
+echo $prog"To use alias, please run \`source $ALIAS_FILE\`."
 echo $prog"You can also add '$INSTALL_DIR' to your PATH variable."
 echo $prog"Installation complete."
